+++ ehrest/src/main/java/com/everhomes/rest/news/SearchNewsCommand.java
@@ -22,7 +22,7 @@ public class SearchNewsCommand {
 	private String ownerType;
 	@NotNull
 	private Long ownerId;
-	private Integer categoryId;
+	private Long categoryId;
 	private String keyword;
 	private Long pageAnchor;
 	private Integer pageSize;
@@ -72,12 +72,12 @@ public class SearchNewsCommand {
 		return StringHelper.toJsonString(this);
 	}
 
-	public Integer getCategoryId() {
+	public Long getCategoryId() {
 		return categoryId;
 	}
 
-	public void setCategoryId(Integer categoryId) {
+	public void setCategoryId(Long categoryId) {
 		this.categoryId = categoryId;
-	}
+	} 
 
 }
